git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@11195 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/renderlib.h
@@ -47,7 +47,7 @@ tRenderCodes RotateVertexList (int nVerts, ushort* vertexIndexP);
 void RotateSideNorms (void);
 // Given a list of point numbers, project any that haven't been projected
 void ProjectVertexList (int nv, ushort *pointIndex);
-ubyte ProjectRenderPoint (short nVertex);
+ubyte ProjectRenderPoint (int nVertex);
 
 void TransformSideCenters (void);
 #if USE_SEGRADS
